backend/requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
psycopg[binary]>=3.2.10
alembic>=1.14.1
pydantic>=2.10.4
pydantic-settings>=2.7.1
python-dotenv>=1.0.1
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
twilio>=9.0.0
python-multipart>=0.0.9
anthropic>=0.40.0
httpx>=0.27.0
